+++ hitt-cli/tests/completions.rs
@@ -3,7 +3,8 @@ mod completions_command {
     use predicates::prelude::PredicateBooleanExt;
 
     fn completions_command() -> assert_cmd::Command {
-        let mut command = assert_cmd::Command::cargo_bin("hitt").unwrap();
+        let mut command =
+            assert_cmd::Command::cargo_bin("hitt").expect("error setting up hitt binary");
 
         command.arg("completions");