+++ CHANGELOG.md
@@ -18,6 +18,7 @@ All notable changes to this project will be documented in this file. Dates are d
 - build(deps): bump once_cell from 1.21.2 to 1.21.3 [`#479`](https://github.com/hougesen/hitt/pull/479)
 - build(deps): bump once_cell from 1.21.1 to 1.21.2 [`#477`](https://github.com/hougesen/hitt/pull/477)
 - build(deps): bump jsonformat from 2.0.0 to 2.1.0 [`#476`](https://github.com/hougesen/hitt/pull/476)
+- test(cli): test completions command [`de78ce7`](https://github.com/hougesen/hitt/commit/de78ce70fcaeada06f82d78d21473a2e17c08a2d)
 
 #### [v0.0.19](https://github.com/hougesen/hitt/compare/v0.0.18...v0.0.19)